step4-hello-agent/shetani-calculator/pyproject.toml
[project]
name = "shetani-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.17",
    "openai>=1.23.0",
    "python-dotenv>=1.1.0",
]